+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Yii1 PSR Log extension
 
 - Enh: Added ability for global log context setup at `PsrLogger` (klimov-paul)
 - Enh: Added `PsrLogger::new()` static method for a new instance creation (klimov-paul)
+- Enh: Added `Logger::$logMessageContextSeparator` allowing to control string separator for log context in Yii log messages (klimov-paul)
 
 
 1.0.0, July 19, 2023